+++ src/main/java/com/sgcc/bg/service/impl/BgNonProjectServiceImpl.java
@@ -56,8 +56,8 @@ public class BgNonProjectServiceImpl implements BgNonProjectService {
 	public List<Map<String, String>> getAllProjects(String proName,String proStatus) {
 		List<Map<String, String>> list=bgNonProjectMapper.getAllProjects(webUtils.getUsername(),proName, proStatus);
 		for(Map<String,String> map : list){
-			Integer hourSum = bgNonProjectMapper.hourSum(map.get("id"));
-			Integer qianQiSum = bgNonProjectMapper.qianQiSum(map.get("id"));
+			Double hourSum = bgNonProjectMapper.hourSum(map.get("id"));
+			Double qianQiSum = bgNonProjectMapper.qianQiSum(map.get("id"));
 			map.put("hourSum", String.valueOf(hourSum+qianQiSum));
 		}
 		return list;
